+++ src/components/Icon/icons/IconCreateFile/index.js
@@ -7,6 +7,8 @@ import image from './create-file.svg';
 const styles = css`
   width: 12px;
   height: 12px;
+  fill: #F5222D;
+  fill-opacity: 0.65;
 `;
 
 export const IconCreateFile = ({ className, onClick }: GenericIconProps) => (