+++ src/components/Icon/icons/IconCreateFolder/index.js
@@ -7,6 +7,8 @@ import image from './create-folder.svg';
 const styles = css`
   width: 12px;
   height: 12px;
+  fill: #F5222D;
+  fill-opacity: 0.65;
 `;
 
 export const IconCreateFolder = ({ className, onClick }: GenericIconProps) => (