+++ src/components/Icon/icons/IconFile/index.js
@@ -1,11 +1,17 @@
 // @flow
 import * as React from 'react';
+import { css, cx } from 'emotion';
 import { Icon, type GenericIconProps } from '../../Icon';
 import image from './file.svg';
 
+const styles = css`
+  fill: #000;
+  fill-opacity: 0.65;
+`;
+
 export const IconFile = ({ className, onClick }: GenericIconProps) => (
   <Icon
-    className={className}
+    className={cx(styles, className)}
     glyph={image}
     onClick={onClick}
   />