+++ src/components/Icon/icons/IconFolder/index.js
@@ -1,8 +1,14 @@
 // @flow
 import * as React from 'react';
+import { css, cx } from 'emotion';
 import { Icon, type GenericIconProps } from '../../Icon';
 import folder from './folder.svg';
 import folderOpened from './folder-opened.svg';
+
+const styles = css`
+  fill: #000;
+  fill-opacity: 0.65;
+`;
 
 type IconFolderProps = {
   ...$Exact<GenericIconProps>,
@@ -11,7 +17,7 @@ type IconFolderProps = {
 
 export const IconFolder = ({ className, onClick, opened }: IconFolderProps) => (
   <Icon
-    className={className}
+    className={cx(styles, className)}
     glyph={opened ? folderOpened : folder}
     onClick={onClick}
   />