+++ src/components/Icon/icons/IconRefresh/index.js
@@ -7,6 +7,8 @@ import image from './refresh.svg';
 const styles = css`
   width: 12px;
   height: 12px;
+  fill: #F5222D;
+  fill-opacity: 0.45;
 `;
 
 export const IconRefresh = ({ className }: GenericIconProps) => (