+++ src/components/Icon/index.js
@@ -31,7 +31,7 @@ export * from './icons/IconOk';
 export * from './icons/IconRadio';
 export * from './icons/IconRefresh';
 export * from './icons/IconSchema';
-export * from './icons/IconSpinner';
 export * from './icons/IconSearch';
+export * from './icons/IconSpinner';
 export * from './icons/IconUser';
 export * from './icons/IconUsers';