+++ src/components/Icon/icons/IconChip/index.js
@@ -1,11 +1,17 @@
 // @flow
 import * as React from 'react';
+import { css, cx } from 'emotion';
 import { Icon, type GenericIconProps } from '../../Icon';
 import image from './chip.svg';
 
+const style = css`
+  fill: #000;
+  fill-opacity: 0.65;
+`;
+
 export const IconChip = ({ className }: GenericIconProps) => (
   <Icon
-    className={className}
+    className={cx(style, className)}
     glyph={image}
   />
 );