+++ src/components/Icon/icons/IconEdit/index.js
@@ -7,6 +7,8 @@ import image from './edit.svg';
 const styles = css`
   width: 12px;
   height: 12px;
+  fill: #F5222D;
+  fill-opacity: 0.65;
 `;
 
 export const IconEdit = ({ className, onClick }: GenericIconProps) => (