+++ src/components/Icon/icons/IconInfo/index.js
@@ -7,6 +7,7 @@ import image from './info.svg';
 const styles = css`
   width: 12px;
   height: 12px;
+  fill: #F5222D;
 `;
 
 export const IconInfo = ({ className }: GenericIconProps) => (