+++ src/components/Icon/icons/IconLink/index.js
@@ -7,6 +7,8 @@ import image from './link.svg';
 const styles = css`
   width: 14px;
   height: 14px;
+  fill: #000;
+  fill-opacity: 0.65;
 `;
 
 export const IconLink = ({ className }: GenericIconProps) => (