+++ src/components/Icon/icons/IconChipWarning/index.js
@@ -1,0 +1,16 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { Icon, type GenericIconProps } from '../../Icon';
+import image from './chipWarning.svg';
+
+const style = css`
+  fill: #FAAD14;
+`;
+
+export const IconChipWarning = ({ className }: GenericIconProps) => (
+  <Icon
+    className={cx(style, className)}
+    glyph={image}
+  />
+);