+++ src/components/Icon/icons/IconAttention/index.js
@@ -2,14 +2,16 @@
 import * as React from 'react';
 import { css, cx } from 'emotion';
 import { Icon, type GenericIconProps } from '../../Icon';
-import image from './bucket.svg';
+import image from './attention.svg';
 
 const styles = css`
+  width: 12px;
+  height: 12px;
   fill: #000;
   fill-opacity: 0.65;
 `;
 
-export const IconBucket = ({ className }: GenericIconProps) => (
+export const IconAttention = ({ className }: GenericIconProps) => (
   <Icon
     className={cx(styles, className)}
     glyph={image}