+++ src/components/Icon/index.js
@@ -1,6 +1,7 @@
 // @flow
 export * from './Icon';
 export * from './icons/IconAttach';
+export * from './icons/IconAttention';
 export * from './icons/IconBox';
 export * from './icons/IconBoxNoData';
 export * from './icons/IconBucket';