+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Updated FlatList styles
 
+- Added syntax highlight in markdown code blocks
+
 ## [0.11.0] - 2020-01-17
 
 - Links in Markdown have a target='\_blank' attribute.