+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Loading Button state looks more disabled than before
 
+- Added icons:
+  - IconChipWarning
+  - IconChipDanger
+
 ## [0.14.0] - 2020-03-20
 
 - Added IconGeoPin