+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Loading Button state looks more disabled than before
 
 - Added icons:
+  - IconAttention
   - IconChipWarning
   - IconChipDanger