+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - IconChipWarning
   - IconChipDanger
 
+- Added `onSubmit` prop to Modal component
+
 ## [0.14.0] - 2020-03-20
 
 - Added IconGeoPin